feat: Added new DTO classes and updated Maven configuration
This commit introduces `Player` and `Tournament` DTO classes along with a `PlayerMapper` to map between entity and DTO. Additionally, Maven configuration is updated to use Java version property for release and the Maven distribution URL is updated to a newer version. The scope of Lombok dependency is also removed from the pom.xml file.